refactor: extract out payment utils
